refactor(worker): extract the Tokio data-plane path into tokio_conn
`main.rs` was 1509 lines spanning CLI parsing, backend construction, io_uring
capability probing, ring supervision, registration, heartbeat, signal handling
*and* the whole Tokio data-plane connection handler.

The connection handler is the part with a natural seam. `uring_conn.rs` was
already a library module, and its own docs describe it as "the completion-based
counterpart to the Tokio `handle_conn` in `main.rs`" — so the two
implementations of one protocol lived at different layers, one in the library
and one inline in a binary. This makes them structurally parallel siblings.

Moved verbatim (bodies are byte-identical; `git diff` shows moves and imports
only): handle_conn, handle_control_frame, handle_put, handle_delete,
sendfile_payload, read_control. Only handle_conn and read_control are `pub`;
the rest stay private so the library's public API does not widen.

main.rs 1509 -> 1059 lines.

Measured effect worth noting: with the path in its own file, coverage now
attributes it separately, and the Tokio fallback sits at 36.4% line coverage
against uring_conn's 78.7%. Both paths must stay behaviourally identical, so
that gap is a real gap — it was previously hidden inside main.rs's aggregate.
Not addressed here; this change only makes it visible.

Refs #306.
